+++ packages/web/blog/src/app/guestbook/components/GuestbookMessage.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import { postGuestbook } from '@apis';
+import { SquareBox } from '@components';
 import { Button, Input, Spacing } from '@guesung/ui';
 import { FormEventHandler, useRef } from 'react';
 
@@ -17,18 +18,20 @@ export default function GuestbookMessage() {
   };
 
   return (
-    <form onSubmit={handleSubmit}>
-      <p>메시지</p>
-      <Spacing size={8} />
-      <div className="flex justify-between gap-4">
-        <Input
-          placeholder="여기에 메시지를 입력해주세요"
-          className="flex-1"
-          ref={inputRef}
-        />
-        <Button>메시지 등록</Button>
-      </div>
-      <Spacing size={24} />
-    </form>
+    <SquareBox>
+      <form onSubmit={handleSubmit}>
+        <p>메시지</p>
+        <Spacing size={8} />
+        <div className="flex justify-between gap-4">
+          <Input
+            placeholder="여기에 메시지를 입력해주세요"
+            className="flex-1"
+            ref={inputRef}
+          />
+          <Button>메시지 등록</Button>
+        </div>
+        <Spacing size={24} />
+      </form>
+    </SquareBox>
   );
 }
