+++ packages/web/blog/src/app/guestbook/components/index.ts
@@ -2,3 +2,4 @@ export { default as GuestbookLogin } from './GuestbookLogin';
 export { default as GuestbookList } from './GuestbookList';
 export { default as GuestbookMessage } from './GuestbookMessage';
 export { default as GuestbookUserInfo } from './GuestbookUserInfo';
+export { default as LoginButton } from './LoginButton';
