+++ packages/web/blog/src/components/Layout.tsx
@@ -15,11 +15,15 @@ export default function Layout({
   );
 }
 
+//
+
 interface TitleProps extends ClassNameType {}
 
 function Title({ children, className }: PropsWithChildren<TitleProps>) {
   return <div className={cn('text-subtitle2', className)}>{children}</div>;
 }
+
+//
 
 interface DescriptionProps extends ClassNameType {}
 
@@ -33,6 +37,8 @@ function Description({
     </div>
   );
 }
+
+//
 
 Layout.Title = Title;
 Layout.Description = Description;
